Removed the TimecardEntry and Image components, took out support for images and resources in general, modified a few editors in behavior, name, and status, re-adjusted the ComponentTypes editor, updated the ribbon, and re-adjusted the EditorTypes editor.

+++ Editors/EditorRegistry.cpp
@@ -95,7 +95,7 @@ namespace Armin::Editors
         return Vector<EditorFrame*>(Info);
     }
 
-    EditorFrame* EditorRegistry::GetEditor(uint EditorType, EditorHost* Parent)
+    EditorFrame* EditorRegistry::GetEditor(ulong EditorType, EditorHost* Parent)
     {
         for (uint i = 0; i < Info.Size; i++)
         {
@@ -106,7 +106,7 @@ namespace Armin::Editors
 
         return nullptr;
     }
-    Vector<EditorFrame*> EditorRegistry::GetEditors(uint EditorType, EditorHost* Parent)
+    Vector<EditorFrame*> EditorRegistry::GetEditors(ulong EditorType, EditorHost* Parent)
     {
         Vector<EditorFrame*> Return;
         for (uint i = 0; i < Info.Size; i++)
@@ -219,7 +219,7 @@ namespace Armin::Editors
             RedrawWindow(*Hosts, NULL, NULL, RDW_INVALIDATE | RDW_ERASENOW);
     }
 
-    void EditorRegistry::ResetEditorOfType(uint Types)
+    void EditorRegistry::ResetEditorOfType(ulong Types)
     {
         for (uint i = 0; i < Info.Size; i++)
         {
@@ -228,7 +228,7 @@ namespace Armin::Editors
                 Current->Reset();
         }
     }
-    void EditorRegistry::ResetEditorOfType(uint Types, Vector<void*> Condition)
+    void EditorRegistry::ResetEditorOfType(ulong Types, Vector<void*> Condition)
     {
         for (uint i = 0; i < Info.Size; i++)
         {
@@ -238,7 +238,7 @@ namespace Armin::Editors
         }
     }
 
-    bool EditorRegistry::IsOpen(uint Types, EditorHost* Host)
+    bool EditorRegistry::IsOpen(ulong Types, EditorHost* Host)
     {
         for (EditorFrame* Frame : Info)
         {
@@ -248,7 +248,7 @@ namespace Armin::Editors
 
         return false;
     }
-    bool EditorRegistry::IsOpen(uint Types, Vector<void*> Args, EditorHost* Host)
+    bool EditorRegistry::IsOpen(ulong Types, Vector<void*> Args, EditorHost* Host)
     {
         for (EditorFrame* Frame : Info)
         {
