+++ Editors/EditorRegistry.h
@@ -40,8 +40,8 @@ namespace Armin
 			static Vector<EditorFrame*> CurrentApplyableEditors();
 			static Vector<EditorFrame*> CurrentOpenEditors();
 
-			static EditorFrame* GetEditor(uint EditorType, EditorHost* Parent = nullptr);
-			static Vector<EditorFrame*> GetEditors(uint EditorType, EditorHost* Parent = nullptr);
+			static EditorFrame* GetEditor(ulong EditorType, EditorHost* Parent = nullptr);
+			static Vector<EditorFrame*> GetEditors(ulong EditorType, EditorHost* Parent = nullptr);
 
 			/// <summary>
 			/// Moves an editor from one editor host to another.
@@ -55,13 +55,13 @@ namespace Armin
 			/// Resets an editor that matches a specified type.
 			/// </summary>
 			/// <param name="Types">The editor types to reset. Can use more than one type of editor.</param>
-			static void ResetEditorOfType(uint Types);
+			static void ResetEditorOfType(ulong Types);
 			/// <summary>
 			/// Resets an editor that matches a specified type and condition.
 			/// </summary>
 			/// <param name="Types">The editor types to reset. Can use more than one type of editor.</param>
 			/// <param name="Condition">The condition the editor must match to be reset.</param>
-			static void ResetEditorOfType(uint Types, Vector<void*> Condition);
+			static void ResetEditorOfType(ulong Types, Vector<void*> Condition);
 
 			/// <summary>
 			/// Evaluates if an editor is open.
@@ -69,7 +69,7 @@ namespace Armin
 			/// <param name="Types">The allowed types of editor to test for.</param>
 			/// <param name="Host">The Host to search under. If nullptr, all hosts are allowed.</param>
 			/// <returns>If the editor matching the specified criteria is open.</returns>
-			static bool IsOpen(uint Types, EditorHost* Host = nullptr);
+			static bool IsOpen(ulong Types, EditorHost* Host = nullptr);
 			/// <summary>
 			/// Evaluates if an editor is open.
 			/// </summary>
@@ -77,7 +77,7 @@ namespace Armin
 			/// <param name="Args">The condition that the editor must match.</param>
 			/// <param name="Host">The Host to search under. If nullptr, all hosts are allowed.</param>
 			/// <returns>If the editor matching the specified criteria is open.</returns>
-			static bool IsOpen(uint Types, Vector<void*> Args, EditorHost* Host = nullptr);
+			static bool IsOpen(ulong Types, Vector<void*> Args, EditorHost* Host = nullptr);
 
 			/// <summary>
 			/// Opens an editor, and specifies it's parent.
